--- a/data/covers80/hparams.yaml
+++ b/data/covers80/hparams.yaml
@@ -6,6 +6,7 @@
   "name": "cqt_with_asr_noise",
   "noise_path": "dataset/asr_as_noise/dataset.txt"
 }
+"device": "mps"
 "train-sample_data_split": 0.1 # percent of training data to reserve for validation aka "train-sample"
 "train-sample_unseen": 0.02 # percent of song_ids from training data to reserve exclusively for validation aka "train-sample"
 "test_data_split": 0.1 # percent of training data to reserve for test aka "dev"