Reduce the data preparation CQT step duration using nnAudio's implementation.

$ time python3 -m tools.extract_csi_features data/covers80_testset
before:
real    1m31,340s
user    17m35,827s
sys     0m10,517s

after (with cuda):
real    0m10,519s
user    1m33,332s
sys     0m10,304s

A new device parameter is added to data/covers80_testset/hparams.yaml to select the device, similar to the train parameter. It is not ideal, but when I tried auto-detection of the device, I faced this error:
> Cannot re-initialize CUDA in forked subprocess. To use CUDA with multiprocessing, you must use the 'spawn' start method
It can be improved later.

The librosa implementation is kept as it is faster when the device is set to "cpu".